requirements.txt
# =============================================================================
# DEEPEMIA PROJECT REQUIREMENTS
# =============================================================================

# Core Python packages
absl-py==0.15.0
typing_extensions==5.4.0

# =============================================================================
# DEEP LEARNING & ML FRAMEWORKS
# =============================================================================

# PyTorch ecosystem (CUDA 11.3 compatible)
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113

# TensorFlow ecosystem
tensorflow==2.6.5
tensorflow-estimator==2.6.0
tensorflow-hub==0.16.1
tensorflow-io==0.21.0
tensorflow-metadata==0.14.0
tensorflow-probability==0.24.0
tf-keras==2.15.0
keras==2.6.0
Keras-Preprocessing==1.1.2

# Computer Vision & Image Processing
detectron2==0.6+cu113
opencv-python-headless==4.11.0.86
Pillow==9.1.0
albumentations==1.3.1
scikit-image==0.19.3
imageio==2.34.0
PyWavelets==1.4.1
tifffile==2024.4.18

# Scientific Computing
numpy==1.23.5
scipy==1.10.1
pandas==1.4.4
matplotlib==3.6.3
scikit-learn==1.4.2
seaborn==0.12.2

# =============================================================================
# GOOGLE CLOUD PLATFORM
# =============================================================================

# Core GCP libraries
google-cloud-storage==2.11.0
google-cloud-core==2.3.2
google-api-core==1.34.1
google-auth==1.35.0
google-cloud-aiplatform==1.41.0
google-cloud-bigquery==3.18.0
google-cloud-vision==3.6.0

# GCP utilities
gcsfs==2024.3.1
google-resumable-media==2.7.0
google-crc32c==1.5.0

# =============================================================================
# WEB INTERFACE & API
# =============================================================================

# Streamlit for GUI (legacy)
streamlit==1.26.0

# Gradio for modern GUI
gradio==4.15.0

# FastAPI for API endpoints
fastapi==0.98.0
uvicorn==0.23.0
starlette==0.27.0

# Web utilities
requests==2.29.0
urllib3==1.26.15

# =============================================================================
# DATA PROCESSING & UTILITIES
# =============================================================================

# Data manipulation
pyarrow==9.0.0
h5py==3.1.0
tqdm==4.65.0

# Configuration & logging
PyYAML==6.0
omegaconf==2.3.0
hydra-core==1.3.2

# OCR and text processing
easyocr==1.6.1

# =============================================================================
# JUPYTER & DEVELOPMENT
# =============================================================================

# Jupyter ecosystem
jupyter==1.0.0
jupyterlab==3.6.6
ipykernel==6.21.3
notebook==6.5.4

# Development tools
black==21.4b2
debugpy==1.6.6

# =============================================================================
# CLOUD & INFRASTRUCTURE
# =============================================================================

# Docker & Kubernetes
docker==7.0.0
kubernetes==29.0.0

# Ray for distributed computing
ray==1.8.0

# Apache Beam for data pipelines
apache-beam==2.46.0

# =============================================================================
# MONITORING & PROFILING
# =============================================================================

# Performance monitoring
gpustat==1.0.0
nvidia-ml-py==11.495.46
Pympler==1.1
psutil==5.9.8

# TensorBoard
tensorboard==2.6.0
tensorboard-data-server==0.6.1

# =============================================================================
# SECURITY & AUTHENTICATION
# =============================================================================

# Authentication
cryptography==40.0.1
PyJWT==2.8.0
oauthlib==3.2.2

# =============================================================================
# ASYNC & NETWORKING
# =============================================================================

# Async libraries
aiohttp==3.8.4
aiofiles==22.1.0
anyio==4.1.0

# Redis
redis==5.0.3
hiredis==2.3.2

# =============================================================================
# UTILITIES & HELPERS
# =============================================================================

# File handling
pathlib2==2.3.7
click==8.1.7
colorama==0.4.6

# Date/time
python-dateutil==2.8.2
pytz==2024.1

# Validation
validators==0.35.0
pydantic==1.9.2

# Progress bars
tqdm==4.65.0

# =============================================================================
# CONDA/MAMBA PACKAGES (if using conda environment)
# =============================================================================

# Note: These packages with file paths are typically conda-installed
# and should be managed through conda, not pip

# conda install -c conda-forge:
# - certifi
# - charset-normalizer  
# - idna
# - jinja2
# - markupsafe
# - packaging
# - setuptools
# - wheel

# =============================================================================
# NVIDIA CUDA LIBRARIES (automatically managed with PyTorch)
# =============================================================================

# These are automatically installed with torch+cu113:
# nvidia-cublas-cu12==12.6.4.1
# nvidia-cuda-runtime-cu12==12.6.77
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cufft-cu12==11.3.0.4
# nvidia-curand-cu12==10.3.7.77
# nvidia-cusolver-cu12==11.7.1.2
# nvidia-cusparse-cu12==12.5.4.2

# =============================================================================
# INSTALLATION NOTES
# =============================================================================

# For CUDA support:
# 1. Install CUDA 11.3 toolkit from NVIDIA
# 2. Ensure compatible GPU drivers
# 3. Install PyTorch with CUDA support first

# For Detectron2:
# pip install detectron2 -f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html

# For development:
# pip install -r requirements.txt --find-links https://download.pytorch.org/whl/cu113/torch_stable.html